refactor(113176): remove arquivos inutilizados
