+++ src/components/screens/Cadastros/EditaisContratos/helper.js
@@ -81,12 +81,6 @@ export const renderizarLabelEmpresa = (selected, options) => {
   return `${selected.length} empresas selecionadas`;
 };
 
-export const montaEstadoEditalEContrato = (edital, contratos) => {
-  const editalContrato = edital;
-  editalContrato["contratos_relacionados"] = contratos;
-  return editalContrato;
-};
-
 export const montaEstadoEditais = (response) => {
   let editais = response.data.results.map((edital) => {
     return {
@@ -100,58 +94,4 @@ export const montaEstadoEditais = (response) => {
     };
   });
   return editais;
-};
-
-export const montaContratoRelacionado = (
-  contratos_relacionados,
-  contrato,
-  indice_contrato
-) => {
-  contratos_relacionados[indice_contrato].lotes = contrato.lotes.map((lote) => {
-    return lote.uuid;
-  });
-  contratos_relacionados[indice_contrato].lotes_nomes = contrato.lotes.map(
-    (lote) => {
-      return lote.nome;
-    }
-  );
-
-  contratos_relacionados[indice_contrato].dres =
-    contrato.diretorias_regionais.map((dre) => {
-      return dre.uuid;
-    });
-  contratos_relacionados[indice_contrato].dres_nomes =
-    contrato.diretorias_regionais.map((dre) => {
-      return dre.nome;
-    });
-
-  contratos_relacionados[indice_contrato].empresas = [
-    contrato.terceirizada.uuid,
-  ];
-
-  contratos_relacionados[indice_contrato].empresas_nomes =
-    contrato.terceirizada.nome_fantasia;
-  contratos_relacionados[indice_contrato].processo_administrativo =
-    contrato.processo;
-  contratos_relacionados[indice_contrato].numero_contrato = contrato.numero;
-  contratos_relacionados[indice_contrato].data_proposta =
-    contrato.data_proposta;
-  contrato.vigencias.forEach((vigencia, indice_vigencia) => {
-    if (indice_vigencia === 0) {
-      contratos_relacionados[indice_contrato].vigencias[
-        indice_vigencia
-      ].data_inicial = vigencia["data_inicial"];
-      contratos_relacionados[indice_contrato].vigencias[
-        indice_vigencia
-      ].data_final = vigencia["data_final"];
-    } else {
-      let vigencia_temp = {
-        data_inicial: vigencia["data_inicial"],
-        data_final: vigencia["data_final"],
-      };
-      contratos_relacionados[indice_contrato].vigencias.push(vigencia_temp);
-    }
-  });
-
-  return contratos_relacionados;
 };
